# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CI4 CMS API
  description: |
    CodeIgniter 4 기반 멀티테넌시 CMS 플랫폼 API 문서입니다.

    ## 특징
    - URL 기반 테넌트 분리
    - Shield RBAC 권한 관리
    - RESTful API 설계
    - Bearer Token 인증

    ## 인증
    대부분의 엔드포인트는 Bearer Token 인증이 필요합니다.
    로그인 후 받은 토큰을 Authorization 헤더에 포함하세요:
    ```
    Authorization: Bearer {your_token}
    ```
  version: 1.0.0
  contact:
    name: CI4 CMS
    url: https://github.com/nambak/ci4-cms
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /posts/{id}:
    get:
      tags:
        - Posts
      summary: 포스트 상세
      description: 특정 포스트의 상세 정보를 조회합니다.
      operationId: getPost
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 포스트 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Post'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Posts
      summary: 포스트 수정
      description: 포스트를 수정합니다. (Editor 이상 권한 필요)
      operationId: updatePost
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '200':
          description: 포스트 수정 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - Posts
      summary: 포스트 삭제
      description: 포스트를 삭제합니다. (Admin 권한 필요)
      operationId: deletePost
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 포스트 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 토큰을 사용한 인증
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: 리소스 ID
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          example: success
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 작업이 성공적으로 완료되었습니다.
    BaseErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: error
        code:
          type: integer
          example: 400
        message:
          type: string
          example: 오류가 발생했습니다.
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseErrorResponse'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: string
              example:
                title: The title field is required.
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: 홍길동
        role:
          type: string
          example: editor
        created_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
...
    Post:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: 포스트 제목
        slug:
          type: string
          example: post-title
        content:
          type: string
          example: 포스트 내용입니다...
        excerpt:
          type: string
          example: 포스트 요약
        status:
          type: string
          enum: [draft, published, archived]
          example: published
        category_id:
          type: integer
          example: 1
        author_id:
          type: integer
          example: 1
        published_at:
          type: string
          format: date-time
          nullable: true
          example: '2025-01-01T00:00:00Z'
        created_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
        category:
          $ref: '#/components/schemas/Category'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        author:
          $ref: '#/components/schemas/User'
    PostInput:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          example: 새로운 포스트
        slug:
          type: string
          example: new-post
        content:
          type: string
          example: 포스트 내용입니다...
        excerpt:
          type: string
          example: 포스트 요약
        status:
          type: string
          enum: [draft, published, archived]
          default: draft
          example: draft
        category_id:
          type: integer
          example: 1
        tags:
          type: array
          items:
            type: integer
          example: [1, 2, 3]

    # Category Schemas
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 기술
        slug:
          type: string
          example: tech
        description:
          type: string
          example: 기술 관련 포스트
        created_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: PHP
        created_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
...
  responses:
...
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BaseErrorResponse'
              - type: object
                properties:
                  code:
                    example: 401
                  message:
                    example: 인증이 필요합니다.
    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BaseErrorResponse'
              - type: object
                properties:
                  code:
                    example: 403
                  message:
                    example: 이 작업을 수행할 권한이 없습니다.
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BaseErrorResponse'
              - type: object
                properties:
                  code:
                    example: 404
                  message:
                    example: 요청한 리소스를 찾을 수 없습니다.
...
    ValidationError:
      description: 유효성 검증 실패
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  code:
                    example: 422
                  message:
                    example: 입력 데이터가 유효하지 않습니다.